Tidy up edit plugins
